ecommerce_api/requirements.txt
asgiref==3.11.0
colorama==0.4.6
Django==6.0
djangorestframework==3.16.1
factory_boy==3.3.3
Faker==40.1.0
inflection==0.5.1
iniconfig==2.3.0
packaging==25.0
pluggy==1.6.0
Pygments==2.19.2
pytest==9.0.2
pytest-django==4.11.1
pytest-factoryboy==2.8.1
python-decouple==3.8
setuptools==80.9.0
sqlparse==0.5.4
typing_extensions==4.15.0
tzdata==2025.2
wheel==0.45.1
